Remove fixme about objref, we don't model enough yet to address the concern
TL;DR we need to the model the objects within the kernel, or come up
with a smart trick (monoticity might help here).

When writing the note, I was confused about between objects, object
refs (addresses of objects), capabilities (object containing object
refs, that is, pointers to objects) and capability pointers (what the
threads have access to).

I was thinking cap = object, and obj_ref were unique identifiers, so
deleting the cap would delete the object(and for example, for an
endpoint, it would drop the list of threads blocked on that endpoint).
But the assumptions are wrong, so the reasoning is wrong.

If a transitions deletes our cap, but then recreates the *exact* same
cap (that is, with the same rights, same badge, and pointing to the
same object), *the object itself hasn't changed* (so for the example
above, the list of threads blocked on that endpoint wouldn't change,
since it lives on the object itself). It's very clear when you say out
loud that "capabilities are just *rights* to resources*. Taking away
your right and then granting you the exam same right is clearly a
no-op.

Or is it? It's not a no-op if you delete the *last* capability to that
object. Suppose out notify system call does this:

    notify(ch)
        delete all caps to endpoint
        // so endpoint itself is deleted, and threads waiting
        // on that endpoint might get woken up to try waiting on the
        // endpoint again. There is no endpoint anymore, so they error
        // out.
        recreate endpoint and restore all caps as it was before

Our spec would accept this behaviour. But clearly it shouldn't. And
right now our projection doesn't model enough to be able describe this
behavior, let alone disallow it. => we need a bigger projection
(the kernel projection should describe the state of the entire kernel,
not just a given thread's perspective of the kernel).

Signed-off-by: Mathieu PATUREL <m.paturel@student.unsw.edu.au>